Major fix (pathfinding fixed and drawing debug line)
Issue was in pathfinding.cs
Was genereating an Ilist and assigning a new pathnode to be equal to a list at position [i] while casting that list as a path node and then adding pathnode to the grid position list list,  (itterating through the wrong list and also adding unnecessary values was what was causing the list initialized but the list count was empty as it wasn't being populated with any actual values.

+++ Assets/Scripts/LevelGrid.cs
@@ -15,7 +15,7 @@ public class LevelGrid : MonoBehaviour
     // these int values will be setup when making connection with level grid so they only need to be defined once
     [SerializeField] private int width;
     [SerializeField] private int height;
-    [SerializeField] private int cellSize;
+    [SerializeField] private float cellSize;
 
     private void Awake()
     {
